backend/requirements.txt
Django
djangorestframework
django-cors-headers
boto3
djangorestframework-simplejwt
dotenv
psycopg[binary]
drf-spectacular
drf-spectacular[sidecar]